backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.20
Werkzeug==2.2.3
Jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.0
cryptography==41.0.3
psycopg2==2.9.9
gunicorn==21.2.0
alembic==1.11.3
email-validator==2.2.0
PyJWT==2.8.0
pytest-cov==4.1.0